+++ x/flags/flags.go
@@ -3,24 +3,26 @@ package flags
 import (
 	"errors"
 	"fmt"
-	"net/url"
 	"os"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	lddynamodb "github.com/launchdarkly/go-server-sdk-dynamodb"
 	ld "gopkg.in/launchdarkly/go-server-sdk.v5"
-	"gopkg.in/launchdarkly/go-server-sdk.v5/ldcomponents"
 )
 
-var (
-	flagsClient            *Client
-	errClientNotConfigured = errors.New("client not configured")
-)
+var flagsClient *Client
 
-const (
-	defaultSDKKeyEnvironmentVariable = "LAUNCHDARKLY_SDK_KEY"
-)
+const defaultSDKKeyEnvironmentVariable = "LAUNCHDARKLY_SDK_KEY"
+
+type FlagName string
+
+type Client struct {
+	sdkKey           string
+	initWait         time.Duration
+	proxyModeConfig  *proxyModeConfig
+	daemonModeConfig *daemonModeConfig
+	wrappedConfig    ld.Config
+	wrappedClient    *ld.LDClient
+}
 
 func Configure(opts ...ConfigOption) error {
 	c, err := NewClient(opts...)
@@ -46,72 +48,6 @@ func GetDefaultClient() (*Client, error) {
 	}
 
 	return flagsClient, nil
-}
-
-type ConfigOption func(c *Client)
-
-func WithSDKKey(key string) ConfigOption {
-	return func(c *Client) {
-		c.sdkKey = key
-	}
-}
-
-func WithInitWait(t time.Duration) ConfigOption {
-	return func(c *Client) {
-		c.initWait = t
-	}
-}
-
-func WithProxyMode(proxyURL *url.URL) ConfigOption {
-	return func(c *Client) {
-		c.proxyModeConfig = &proxyModeConfig{
-			relayProxyURL: proxyURL.String(),
-		}
-	}
-}
-
-func WithDaemonMode(dynamoTableName string, cacheTTL time.Duration) ConfigOption {
-	return func(c *Client) {
-		if c.daemonModeConfig == nil {
-			c.daemonModeConfig = &daemonModeConfig{}
-		}
-
-		c.daemonModeConfig = &daemonModeConfig{
-			dynamoTableName: dynamoTableName,
-			cacheTTL:        cacheTTL,
-		}
-	}
-}
-
-func WithDynamoBaseURL(baseURL *url.URL) ConfigOption {
-	return func(c *Client) {
-		if c.daemonModeConfig == nil {
-			c.daemonModeConfig = &daemonModeConfig{}
-		}
-
-		c.daemonModeConfig.dynamoBaseURL = baseURL.String()
-	}
-}
-
-type FlagName string
-
-type Client struct {
-	sdkKey           string
-	initWait         time.Duration
-	proxyModeConfig  *proxyModeConfig
-	daemonModeConfig *daemonModeConfig
-	wrappedConfig    ld.Config
-	wrappedClient    *ld.LDClient
-}
-
-type proxyModeConfig struct {
-	relayProxyURL string
-}
-
-type daemonModeConfig struct {
-	dynamoTableName string
-	dynamoBaseURL   string
-	cacheTTL        time.Duration
 }
 
 func NewClient(opts ...ConfigOption) (*Client, error) {
@@ -154,27 +90,6 @@ func (c *Client) Connect() error {
 	flagsClient.wrappedClient = wrappedClient
 
 	return nil
-}
-
-func configForProxyMode(cfg *proxyModeConfig) ld.Config {
-	return ld.Config{
-		DataSource: ldcomponents.StreamingDataSource().BaseURI(cfg.relayProxyURL),
-	}
-}
-
-func configForDaemonMode(cfg *daemonModeConfig) ld.Config {
-	datastoreBuilder := lddynamodb.DataStore(cfg.dynamoTableName)
-
-	if cfg.dynamoBaseURL != "" {
-		datastoreBuilder.ClientConfig(aws.NewConfig().WithEndpoint(cfg.dynamoBaseURL))
-	}
-
-	return ld.Config{
-		DataSource: ldcomponents.ExternalUpdatesOnly(),
-		DataStore: ldcomponents.PersistentDataStore(
-			datastoreBuilder,
-		).CacheTime(cfg.cacheTTL),
-	}
 }
 
 func (c *Client) QueryBool(key FlagName, user User, defaultValue bool) (bool, error) {
